pyproject.toml
[project]
name = "bikeclf"
version = "0.1.0"
description = "Bike relevance classification for German civic reports using Gemini 2.0 Flash"
requires-python = ">=3.10"
authors = [
    {name = "Phase 1 Team"}
]
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "langfuse>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bikeclf", "bikeclf.phase1"]
